moved some things around to be better organized
