+++ app/javascript/psulib_blacklight/javascripts/align_rtl_index.js
@@ -1,0 +1,13 @@
+export const alignRtl = function () {
+  // Align rtl text properly
+  $('article').each(function () {
+    const titleLink = this.querySelector('.index_title').querySelector('a');
+    titleLink.setAttribute('dir', 'auto');
+    if (window.getComputedStyle(titleLink).direction === 'rtl') {
+      titleLink.setAttribute(
+        'class',
+        'float-right text-align-start col-sm-11 p-0 pr-4'
+      );
+    }
+  });
+};
