+++ app/javascript/psulib_blacklight/javascripts/index.js
@@ -1,9 +1,9 @@
-import { alignRtl, linkHighlight } from './show_default.js';
+import './show_default';
+import { alignRtl } from './align_rtl_index.js';
 import './google_preview';
 import './a11y_advanced_search';
 import './bookmark_all_on_page';
 
-$(document).ready(linkHighlight);
 $(document).ready(alignRtl);
 
 // Otherwise bootstrap-select won't fire on turbolinked clicks to Advance Search
