+++ app/javascript/psulib_blacklight/javascripts/show_default.js
@@ -1,4 +1,4 @@
-export const linkHighlight = function () {
+$(document).ready(() => {
   // link highlighting of hierarchy
   $('.search-subject').hover(
     function () {
@@ -8,18 +8,4 @@ export const linkHighlight = function () {
       $(this).prevAll().removeClass('field-hierarchy');
     }
   );
-};
-
-export const alignRtl = function () {
-  // Align rtl text properly
-  $('article').each(function () {
-    const titleLink = this.querySelector('.index_title').querySelector('a');
-    titleLink.setAttribute('dir', 'auto');
-    if (window.getComputedStyle(titleLink).direction === 'rtl') {
-      titleLink.setAttribute(
-        'class',
-        'float-right text-align-start col-sm-11 p-0 pr-4'
-      );
-    }
-  });
-};
+});
